initial set up for switch to ResourcefulConfig lib
Renamed temp class ResourcefulConfig.java to BUConfig and removed old BUConfig. Refactored some methods from new BUConfig into util class.
