+++ src/main/java/com/github/mkram17/bazaarutils/config/BUModMenu.java
@@ -1,11 +1,12 @@
 package com.github.mkram17.bazaarutils.config;
 
+import com.github.mkram17.bazaarutils.config.util.ConfigUtil;
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 
 public class BUModMenu implements ModMenuApi {
     @Override
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return ResourcefulConfig::createGUI;
+        return ConfigUtil::createGUI;
     }
 }
